+++ extensions/sourcegraph/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+- **ai**: This extension now supports Raycast AI with a set of tools that grants Raycast AI access to Sourcegraph code search across your private code (if configured) and the public code search corpus on [Sourcegraph.com](https://sourcegraph.com). To get started, include `@sourcegraph` in your next Raycast AI chat session!
+
 ## [Branding updates and workspaces](https://github.com/raycast/extensions/pull/16793) - 2025-02-03
 
 - **search**: We've updated colors throughout the extension to match Sourcegraph's refreshed branding.